test: split daemon integration root surfaces into modules
